[client] Move TieringSplitGenerator to fluss-client tiering package
Extract engine-agnostic split generation from fluss-flink so that other
engines (e.g. Spark tiering) can reuse it. Replace FlinkRuntimeException
with FlussRuntimeException to drop the Flink dependency.
