+++ fluss-client/src/main/java/org/apache/fluss/client/tiering/TieringSplitGenerator.java
@@ -15,16 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.fluss.flink.tiering.source.split;
+package org.apache.fluss.client.tiering;
 
 import org.apache.fluss.client.admin.Admin;
 import org.apache.fluss.client.initializer.BucketOffsetsRetrieverImpl;
 import org.apache.fluss.client.initializer.OffsetsInitializer.BucketOffsetsRetriever;
 import org.apache.fluss.client.metadata.KvSnapshots;
 import org.apache.fluss.client.metadata.LakeSnapshot;
-import org.apache.fluss.client.tiering.TieringLogSplit;
-import org.apache.fluss.client.tiering.TieringSnapshotSplit;
-import org.apache.fluss.client.tiering.TieringSplit;
+import org.apache.fluss.exception.FlussRuntimeException;
 import org.apache.fluss.exception.LakeTableSnapshotNotExistException;
 import org.apache.fluss.metadata.PartitionInfo;
 import org.apache.fluss.metadata.TableBucket;
@@ -32,7 +30,6 @@ import org.apache.fluss.metadata.TableInfo;
 import org.apache.fluss.metadata.TablePath;
 import org.apache.fluss.utils.ExceptionUtils;
 
-import org.apache.flink.util.FlinkRuntimeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -78,7 +75,7 @@ public class TieringSplitGenerator {
             if (t instanceof LakeTableSnapshotNotExistException) {
                 lakeSnapshotInfo = null;
             } else {
-                throw new FlinkRuntimeException(
+                throw new FlussRuntimeException(
                         String.format(
                                 "Failed to get table snapshot for table %s",
                                 tableInfo.getTablePath()),
@@ -130,7 +127,7 @@ public class TieringSplitGenerator {
                                     .getLatestKvSnapshots(tableInfo.getTablePath(), partitionName)
                                     .get();
                 } catch (Exception e) {
-                    throw new FlinkRuntimeException(
+                    throw new FlussRuntimeException(
                             String.format(
                                     "Failed to get table snapshot for table %s and partition %s",
                                     tableInfo.getTablePath(), partitionName),
@@ -166,7 +163,7 @@ public class TieringSplitGenerator {
             try {
                 latestKvSnapshots = flussAdmin.getLatestKvSnapshots(tableInfo.getTablePath()).get();
             } catch (Exception e) {
-                throw new FlinkRuntimeException(
+                throw new FlussRuntimeException(
                         String.format(
                                 "Failed to get table snapshot for table %s",
                                 tableInfo.getTablePath()),
